Enhancements for sanitizing URLs.
Added `img` `src` tag attribute to the built-in sanitization as it is a publicly supported attribute and contains URI payloads; the same as what we do for `a` `href`.

Broke out the URL sanitization into a public method to make it possible to sanitize URLs according to the rules without necessarily needing to extend the whitelist and/or define custom filtering options to sanitize tag attributes containing URI values.