Adds support for the HTML source tag.
ArcGIS Online does not support the src attribute in the audio or video tags, and instead requires use of the nested source tag within the audio or video tag. This commit updates the whitelist to remove the src attribute from the audio and video tags, and adds support for the source tag.