Rollback xss from 1.0.14 to 1.0.13.
Clients who use `stripIgnoreTag: true` in their configuration see a breaking change in the 3.0.0 release. Prior to upgrading to 1.0.14 using `stripIgnoreTag: true`, clients expected a string like "Wheat < 5 lbs" to be sanitized to "Wheat &lt; 5 lbs". Instead they are seeing "Wheat ". Notably, the ArcGIS API for JavaScript uses `stripIgnoreTag: true` in their configuration, so this is causing significant regression in that API.